refactor: `Player`를 `Lottos`로 대체, `Money`는 따로 구현
- `Player` 삭제
- `Lottos` 일급 컬렉션 생성 및 controller에 적용
